pyproject.toml
[tool.poetry]
name = "squad-awards"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
django = "^4.1.4"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.2.0"
gunicorn = {extras = ["gevent"], version = "^20.1.0"}
requests = "^2.28.1"
dj-database-url = "^1.0.0"
loguru = "^0.6.0"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
